+++ Setting-Abramov-Pavel-iOs-5-HW13/ViewController.swift
@@ -7,31 +7,6 @@
 
 import UIKit
 
-struct Section {
-    let title: String
-    let options: [SettingOptionType]
-}
-
-enum SettingOptionType {
-    case staticCell(model: SettingOption)
-    case switchCell(model: SettingsSwitchOption)
-}
-
-struct SettingsSwitchOption {
-    let title: String
-    let icon: UIImage?
-    let iconBackgroundColor: UIColor
-    let handler: (() -> Void)
-    var isOn: Bool
-}
-
-struct SettingOption {
-    let title: String
-    let icon: UIImage?
-    let iconBackgroundColor: UIColor
-    let handler: (() -> Void)
-    
-}
 
 class ViewController: UIViewController, UITableViewDelegate, UITableViewDataSource {
     
@@ -44,11 +19,11 @@ class ViewController: UIViewController, UITableViewDelegate, UITableViewDataSour
         return table
     }()
     
-    var models = [Section]()
+    private var model = [Section]()
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        configure()
+        model = Model.configure()
         title = "Setting"
         view.addSubview(tableView)
         tableView.delegate = self
@@ -56,84 +31,22 @@ class ViewController: UIViewController, UITableViewDelegate, UITableViewDataSour
         tableView.frame = view.bounds
     }
     
-    func configure() {
-        models.append(Section(title: "", options: [
-            .switchCell(model: SettingsSwitchOption(title: "Airplane Mode", icon: UIImage(systemName: "airplane.circle"), iconBackgroundColor: .red , handler: {
-                        print("Airplane Mode")
-                    }, isOn: true )),
-            
-        ]))
-        
-        models.append(Section(title: "General", options: [
-            .staticCell(model: SettingOption(title: "Wi-Fi", icon: UIImage(systemName: "speaker.fill"), iconBackgroundColor:
-                    .systemPink) {
-                        print("Wi-Fi")
-                    }),
-            .staticCell(model: SettingOption(title: "Bluetooth", icon: UIImage(systemName: "shuffle"), iconBackgroundColor:
-                    .systemBlue) {
-                        print("Bluetooth")
-                    }),
-            .staticCell(model:SettingOption(title: "Cellular communication", icon: UIImage(systemName: "phone.connection"), iconBackgroundColor:
-                    .systemPurple) {
-                        print("Cellular communication")
-                    })
-            
-        ]))
-        
-        models.append(Section(title: "", options: [
-            .switchCell(model:SettingsSwitchOption(title: "Modem mode", icon: UIImage(systemName: "personalhotspot"), iconBackgroundColor:
-                    .systemGreen, handler: {
-                        print("Airplane Mode")
-                    }, isOn: true )),
-            
-        ]))
-        
-        models.append(Section(title: "Information", options: [
-            .staticCell(model: SettingOption(title: "Notifications", icon: UIImage(systemName: "speaker.fill"), iconBackgroundColor: .systemPink) {
-                print("Notifications")
-            }),
-            .staticCell(model: SettingOption(title: "Sounds screen time", icon: UIImage(systemName: "bell"), iconBackgroundColor: .systemRed) {
-                print("Sounds screen time")
-            }),
-            .staticCell(model:SettingOption(title: "Focusing", icon: UIImage(systemName: "bolt"), iconBackgroundColor: .systemPurple) {
-                
-            }),
-            .staticCell(model:SettingOption(title: "Screen time", icon: UIImage(systemName: "timer"), iconBackgroundColor: .systemPurple) {
-                
-            })
-        ]))
-        
-        models.append(Section(title: "Basic", options: [
-            .staticCell(model: SettingOption(title: "Basic", icon: UIImage(systemName: "brain"), iconBackgroundColor: .systemGray) {
-                print("Basic")
-            }),
-            .staticCell(model: SettingOption(title: "Control point", icon: UIImage(systemName: "control"), iconBackgroundColor: .systemGray) {
-                print("Control point")
-            }),
-            .staticCell(model:SettingOption(title: "Screen and brightness", icon: UIImage(systemName: "bell"), iconBackgroundColor: .systemBlue) {
-                
-            }),
-            .staticCell(model:SettingOption(title: "Wallpaper", icon: UIImage(systemName: "square.and.arrow.down"), iconBackgroundColor: .systemPurple) {
-                
-            })
-        ]))
-    }
     
     func tableView(_ tableView: UITableView, titleForHeaderInSection section: Int ) -> String?  {
-        let section = models[section]
+        let section = model[section]
         return section.title
     }
     
     func numberOfSections(in tableView: UITableView) -> Int {
-        return models.count
+        return model.count
     }
     
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return models[section].options.count
+        return model[section].options.count
     }
     
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let model = models[indexPath.section].options[indexPath.row]
+        let model = model[indexPath.section].options[indexPath.row]
         
         switch model.self {
         case .staticCell(let model):
@@ -160,7 +73,7 @@ class ViewController: UIViewController, UITableViewDelegate, UITableViewDataSour
     
     func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
         tableView.deselectRow(at: indexPath, animated: true)
-        let type = models[indexPath.section].options[indexPath.row]
+        let type = model[indexPath.section].options[indexPath.row]
         switch type.self {
         case .staticCell(let model):
             model.handler()
